requirements.txt
gspread==5.12.4
oauth2client==4.1.3
pandas==2.2.0
lightgbm==4.3.0
scikit-learn==1.4.0
joblib==1.3.2
streamlit==1.31.1
plotly==5.18.0 
yagmail
flask==2.3.2
google-api-python-client==2.150.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
matplotlib==3.10.0